backend/pyproject.toml
[project]
name = "shangtugongfang-backend"
version = "0.1.0"
description = "尚图工坊后端服务"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.1",
  "asyncpg>=0.30.0",
  "celery>=5.5.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.8",
  "httpx>=0.28.1",
  "openai>=1.108.0",
  "passlib[bcrypt]>=1.7.4",
  "psycopg[binary]>=3.2.6",
  "pydantic-settings>=2.8.1",
  "pyjwt>=2.10.1",
  "python-multipart>=0.0.20",
  "cos-python-sdk-v5>=1.9.38",
  "redis>=5.2.1",
  "sqlalchemy>=2.0.38",
  "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
  "aiosqlite>=0.21.0",
  "pytest>=8.3.5",
  "pytest-asyncio>=0.25.3",
  "pytest-cov>=6.0.0",
  "ruff>=0.11.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
